requirements.txt
Django==4.2
paho-mqtt==2.0.0
Pillow==10.2.0